+++ src/Kiota.Builder/Writers/Python/CodePropertyWriter.cs
@@ -44,7 +44,16 @@ public class CodePropertyWriter : BaseElementWriter<CodeProperty, PythonConventi
                 var defaultValue = isNonNullableCollection ? "[]" : "None";
                 if (!string.IsNullOrEmpty(codeElement.DefaultValue))
                 {
-                    defaultValue = codeElement.DefaultValue;
+                    if (codeElement.Type is CodeType propertyCodeType && propertyCodeType.TypeDefinition is CodeEnum enumDefinition)
+                    {
+                        var enumTypeName = conventions.GetTypeString(codeElement.Type, codeElement);
+                        var enumValue = codeElement.DefaultValue.Trim('"').CleanupSymbolName().ToFirstCharacterUpperCase();
+                        defaultValue = $"{enumTypeName}.{enumValue}";
+                    }
+                    else
+                    {
+                        defaultValue = codeElement.DefaultValue;
+                    }
                 }
                 writer.WriteLine($"{conventions.GetAccessModifier(codeElement.Access)}{codeElement.NamePrefix}{codeElement.Name}: {(codeElement.Type.IsNullable ? "Optional[" : string.Empty)}{returnType}{(codeElement.Type.IsNullable ? "]" : string.Empty)} = {defaultValue}");
                 writer.WriteLine();