+++ tests/Kiota.Builder.Tests/Writers/Python/CodePropertyWriterTests.cs
@@ -149,4 +149,86 @@ public sealed class CodePropertyWriterTests : IDisposable
         var result = tw.ToString();
         Assert.Contains("return '' if self.error.message is None else self.error.message", result);
     }
+    [Fact]
+    public void WritesQueryParameterWithDefaultEnumValue()
+    {
+        var enumProperty = new CodeProperty
+        {
+            Name = "EnumProperty",
+            Type = new CodeType
+            {
+                Name = "SomeEnum",
+                TypeDefinition = new CodeEnum
+                {
+                    Name = "SomeEnum"
+                }
+            },
+            Kind = CodePropertyKind.QueryParameter,
+            DefaultValue = "EnumValue2",
+        };
+        parentClass.AddProperty(enumProperty, new()
+        {
+            Name = "queryParameter",
+            Kind = CodePropertyKind.QueryParameter,
+            Type = new CodeType
+            {
+                Name = "QueryParameter",
+            },
+        });
+        writer.Write(enumProperty);
+        var result = tw.ToString();
+        Assert.Contains("EnumProperty: Optional[SomeEnum] = SomeEnum.EnumValue2", result);
+    }
+    [Fact]
+    public void WritesQueryParameterWithDefaultIntValue()
+    {
+        var integerProperty = new CodeProperty
+        {
+            Name = "IntegerProperty",
+            Type = new CodeType
+            {
+                Name = "int"
+            },
+            Kind = CodePropertyKind.QueryParameter,
+            DefaultValue = "3",
+        };
+        parentClass.AddProperty(integerProperty, new()
+        {
+            Name = "queryParameter",
+            Kind = CodePropertyKind.QueryParameter,
+            Type = new CodeType
+            {
+                Name = "QueryParameter",
+            },
+        });
+        writer.Write(integerProperty);
+        var result = tw.ToString();
+        Assert.Contains("IntegerProperty: Optional[int] = 3", result);
+    }
+    [Fact]
+    public void WritesQueryParameterWithDefaultStringValue()
+    {
+        var stringProperty = new CodeProperty
+        {
+            Name = "StringProperty",
+            Type = new CodeType
+            {
+                Name = "str"
+            },
+            Kind = CodePropertyKind.QueryParameter,
+            DefaultValue = "\"SomeString\"",
+        };
+        parentClass.AddProperty(stringProperty, new()
+        {
+            Name = "queryParameter",
+            Kind = CodePropertyKind.QueryParameter,
+            Type = new CodeType
+            {
+                Name = "QueryParameter",
+            },
+        });
+        writer.Write(stringProperty);
+        var result = tw.ToString();
+        Assert.Contains("StringProperty: Optional[str] = \"SomeString\"", result);
+    }
 }